requirements.txt
stripe~=11.1.0
Faker~=25.2.0
requests~=2.31.0
razorpay~=1.4.1
hiredis==2.0.0
    # via aioredis
msgpack==1.0.3
    # via
    #   cachecontrol
    #   channels-redis
channels==4.0.0
    # via channels-redis
channels-redis==4.2.0
    # via channels
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   daphne
aioredis==1.3.1
    # via channels-redis
redis==5.0.1
    # via channels-redis
pytz~=2023.3.post1
namesgenerator~=0.3
autobahn==23.6.2
    # via daphne
daphne==4.0.0
    # via channels
twisted==23.10.0
    # via daphne
Django~=4.2.10
celery~=5.3.6
djangorestframework~=3.14.0
arrow~=1.3.0
weasyprint~=62.2
whatsmate~=1.3
drfaddons~=0.1.4
django-unfold==0.39.0